feat : 운동인증 수정 API 추가

+++ src/main/java/fithub/app/service/impl/ArticleServiceImpl.java
@@ -169,9 +169,11 @@ public class ArticleServiceImpl implements ArticleService {
             articleHashTagRepository.delete(articleHashTag);
         }
 
-        ArticleHashTag last = articleHashTagList.get(0);
-        articleHashTagList.remove(last);
-        articleHashTagRepository.delete(last);
+        if (articleHashTagList.size() > 0) {
+            ArticleHashTag last = articleHashTagList.get(0);
+            articleHashTagList.remove(last);
+            articleHashTagRepository.delete(last);
+        }
 
         String exerciseTag =  request.getExerciseTag();
         HashTag exercisehashTag = hashTagRepository.findByName('#' + exerciseTag).orElseGet(() -> HashTagConverter.newHashTag(exerciseTag));
